refactor[backend]: удалено старое обслуживание S3
