+++ docs/superpowers/plans/2026-08-06-timeweb-s3-migration.md
@@ -346,9 +346,18 @@ AI-отчёты сохраняются бессрочно по ключу `org-{
 - [x] **Step 2: Проверить RED**
 - [x] **Step 3: Удалить параметр бакета и динамические S3-диски из нормативных источников**
 - [x] **Step 4: Проверить GREEN и обратные вызовы импорта/обновления ФГИС ЦС**
-- [ ] **Step 5: Commit, PR, merge и deploy**
+- [x] **Step 5: Commit, PR, merge и deploy**
+
+Реализовано в PR #244 (`494f218a9f505447b6b724e6b9e74069eee0ca55`), штатный deploy `31060956039` завершён успешно. Production SHA совпал, профильных ошибок в последних 500 строках журнала нет.
 
 ##### Task 3B.2c: Оставшиеся доменные вызовы
+
+**Подблок 3B.2c.1 — старое обслуживание бакетов и автоудаление отчётов (`refactor/timeweb-s3-domain-callers-v3`):**
+
+- [x] удалить команды сканирования прежних бакетов и общего orphan-cleanup;
+- [x] снять с расписания синхронизацию бакетов и автоматическое удаление отчётов;
+- [x] сохранить отдельную проверку действующей очистки битых аватаров;
+- [ ] commit, PR, merge и deploy.
 
 - [ ] **Step 1: Написать failing архитектурные и доменные тесты**
 - [ ] **Step 2: Проверить RED**
